Backend split into multiple files
