+++ dev/server/package.json
@@ -3,7 +3,8 @@
   "module": "src/index.ts",
   "type": "module",
   "devDependencies": {
-    "@types/bun": "latest"
+    "@types/bun": "^1.1.6",
+    "bun-types": "^1.1.20"
   },
   "peerDependencies": {
     "typescript": "^5.5.3"
