+++ dev/server/src/express/static.ts
@@ -1,0 +1,17 @@
+import express from "express";
+import type { Express } from "express";
+import { mkdir } from "fs/promises";
+import { IMAGES_PATH, WEBSITE_PATH } from "../env";
+
+const SITE_ROUTES = ["/", "/auth", "/ingredients", "/recipe"];
+
+function init(app: Express) {
+  mkdir(IMAGES_PATH, { recursive: true })
+    .catch(console.error)
+    .then(() => {
+      SITE_ROUTES.map((route) => app.use(express.static(WEBSITE_PATH)));
+      app.use("/images", express.static(IMAGES_PATH));
+    });
+}
+
+export { init };
