+++ dev/server/src/model.ts
@@ -104,7 +104,7 @@ UserSchema.method("setPassword", async function (plainPassword: string) {
     let encryptedPassword = await Bun.password.hash(plainPassword);
     this.password = encryptedPassword;
   } catch (error) {
-    console.log("Invalid password, can't set password");
+    console.log("can't set password");
   }
 });
 
@@ -170,7 +170,7 @@ UserSchema.static(
   }): Promise<User | string | null> {
     let existing = await this.findOne({ email: auth.email });
     if (existing) {
-      return "Email already exists";
+      return "email already exists";
     }
     let user = new Users();
     let ingredients = new IngredientsLists({ list: [] });
